--- a/.bazelrc
+++ b/.bazelrc
@@ -33,6 +33,8 @@ build:remote-shared --remote_download_minimal
 build:remote-shared --experimental_repo_remote_exec
 build:remote-shared --jobs=100
 build:remote-shared --verbose_failures
+# Unresolved symlinks are used for rules_js.
+build:remote-shared --experimental_allow_unresolved_symlinks
 
 # Build with --config=remote to use BuildBuddy RBE.
 build:remote --config=remote-shared
@@ -122,6 +124,8 @@ build --incompatible_strict_action_env
 
 # rules_nodejs needs runfiles to be explicitly enabled.
 build --enable_runfiles
+
+build --strategy=TsProject=worker
 
 # Don't build or run Docker tests by default, to avoid a hard dependency
 # on Docker for development.