--- a/.npmrc
+++ b/.npmrc
@@ -1,0 +1,2 @@
+strict-peer-dependencies=false
+