--- a/app/BUILD.bazel
+++ b/app/BUILD.bazel
@@ -1,5 +1,5 @@
+load("@aspect_rules_esbuild//esbuild:defs.bzl", "esbuild")
 load("//rules/typescript:index.bzl", "ts_library")
-load("@npm//@bazel/esbuild:index.bzl", "esbuild")
 load("//rules/sha:index.bzl", "sha")
 
 package(default_visibility = ["//visibility:public"])
@@ -13,7 +13,7 @@ config_setting(
 
 esbuild(
     name = "app_bundle",
-    args = {
+    config = {
         "resolveExtensions": [
             ".mjs",
             ".js",
@@ -35,11 +35,11 @@ ts_library(
     name = "app",
     srcs = glob(["*.tsx"]),
     deps = [
+        "//:node_modules/@types/react",
+        "//:node_modules/@types/react-dom",
+        "//:node_modules/react",
+        "//:node_modules/react-dom",
         "//app/root",
-        "@npm//@types/react",
-        "@npm//@types/react-dom",
-        "@npm//react",
-        "@npm//react-dom",
     ],
 )