--- a/app/components/button/BUILD
+++ b/app/components/button/BUILD
@@ -12,10 +12,10 @@ ts_library(
     ]),
     strict = True,
     deps = [
+        "//:node_modules/@types/react",
+        "//:node_modules/react",
+        "//:node_modules/tslib",
         "//app/components/checkbox",
         "//app/components/link",
-        "@npm//@types/react",
-        "@npm//react",
-        "@npm//tslib",
     ],
 )