--- a/app/components/dialog/BUILD
+++ b/app/components/dialog/BUILD
@@ -9,9 +9,9 @@ ts_library(
     srcs = ["dialog.tsx"],
     strict = True,
     deps = [
-        "@npm//@types/react",
-        "@npm//react",
-        "@npm//tslib",
+        "//:node_modules/@types/react",
+        "//:node_modules/react",
+        "//:node_modules/tslib",
     ],
 )
 
@@ -21,11 +21,11 @@ ts_library(
     strict = True,
     deps = [
         ":dialog",
+        "//:node_modules/@types/react",
+        "//:node_modules/react",
+        "//:node_modules/tslib",
         "//app/components/button",
         "//app/components/modal",
         "//app/components/spinner",
-        "@npm//@types/react",
-        "@npm//react",
-        "@npm//tslib",
     ],
 )