--- a/app/components/digest/BUILD
+++ b/app/components/digest/BUILD
@@ -12,11 +12,11 @@ ts_library(
     ]),
     strict = True,
     deps = [
+        "//:node_modules/@types/long",
+        "//:node_modules/@types/react",
+        "//:node_modules/long",
+        "//:node_modules/react",
+        "//:node_modules/tslib",
         "//app/format",
-        "@npm//@types/long",
-        "@npm//@types/react",
-        "@npm//long",
-        "@npm//react",
-        "@npm//tslib",
     ],
 )