--- a/app/components/filter_input/BUILD
+++ b/app/components/filter_input/BUILD
@@ -12,9 +12,9 @@ ts_library(
     ]),
     strict = True,
     deps = [
-        "@npm//@types/react",
-        "@npm//lucide-react",
-        "@npm//react",
-        "@npm//tslib",
+        "//:node_modules/@types/react",
+        "//:node_modules/lucide-react",
+        "//:node_modules/react",
+        "//:node_modules/tslib",
     ],
 )