--- a/app/components/radio/BUILD
+++ b/app/components/radio/BUILD
@@ -9,8 +9,8 @@ ts_library(
     srcs = glob(["*.tsx"]),
     strict = True,
     deps = [
-        "@npm//@types/react",
-        "@npm//react",
-        "@npm//tslib",
+        "//:node_modules/@types/react",
+        "//:node_modules/react",
+        "//:node_modules/tslib",
     ],
 )