--- a/app/docs/BUILD
+++ b/app/docs/BUILD
@@ -6,6 +6,9 @@ ts_library(
     name = "docs",
     srcs = glob(["*.tsx"]),
     deps = [
+        "//:node_modules/@types/react",
+        "//:node_modules/protobufjs",
+        "//:node_modules/react",
         "//app/auth",
         "//app/capabilities",
         "//app/components/button",
@@ -13,8 +16,5 @@ ts_library(
         "//app/router",
         "//app/service",
         "//proto:bazel_config_ts_proto",
-        "@npm//@types/react",
-        "@npm//protobufjs",
-        "@npm//react",
     ],
 )