--- a/app/footer/BUILD
+++ b/app/footer/BUILD
@@ -7,8 +7,8 @@ ts_library(
     srcs = glob(["*.tsx"]),
     strict = True,
     deps = [
+        "//:node_modules/@types/react",
+        "//:node_modules/react",
         "//app/capabilities",
-        "@npm//@types/react",
-        "@npm//react",
     ],
 )