--- a/app/format/BUILD
+++ b/app/format/BUILD
@@ -1,5 +1,4 @@
 load("//rules/typescript:index.bzl", "ts_jasmine_node_test", "ts_library")
-load("@npm//@bazel/jasmine:index.bzl", "jasmine_node_test")
 
 package(default_visibility = ["//visibility:public"])
 
@@ -7,12 +6,11 @@ ts_library(
     name = "format",
     srcs = ["format.tsx"],
     deps = [
-        "@npm//@types/long",
-        "@npm//@types/moment",
-        "@npm//date-fns",
-        "@npm//long",
-        "@npm//moment",
-        "@npm//tslib",
+        "//:node_modules/@types/long",
+        "//:node_modules/date-fns",
+        "//:node_modules/long",
+        "//:node_modules/moment",
+        "//:node_modules/tslib",
     ],
 )
 
@@ -21,8 +19,8 @@ ts_jasmine_node_test(
     srcs = ["format_test.ts"],
     deps = [
         ":format",
-        "@npm//@types/long",
-        "@npm//long",
-        "@npm//tslib",
+        "//:node_modules/@types/long",
+        "//:node_modules/long",
+        "//:node_modules/tslib",
     ],
 )