--- a/app/menu/BUILD
+++ b/app/menu/BUILD
@@ -6,6 +6,10 @@ ts_library(
     name = "menu",
     srcs = glob(["*.tsx"]),
     deps = [
+        "//:node_modules/@types/react",
+        "//:node_modules/lucide-react",
+        "//:node_modules/protobufjs",
+        "//:node_modules/react",
         "//app/auth",
         "//app/capabilities",
         "//app/preferences",
@@ -13,9 +17,5 @@ ts_library(
         "//app/service",
         "//proto:invocation_ts_proto",
         "//proto:user_ts_proto",
-        "@npm//@types/react",
-        "@npm//lucide-react",
-        "@npm//protobufjs",
-        "@npm//react",
     ],
 )