--- a/app/root/BUILD.bazel
+++ b/app/root/BUILD.bazel
@@ -9,6 +9,9 @@ ts_library(
     srcs = glob(["*.tsx"]),
     strict = True,
     deps = [
+        "//:node_modules/@types/react",
+        "//:node_modules/protobufjs",
+        "//:node_modules/react",
         "//app/alert",
         "//app/auth",
         "//app/capabilities",
@@ -23,8 +26,5 @@ ts_library(
         "//app/router",
         "//app/service",
         "//proto:user_ts_proto",
-        "@npm//@types/react",
-        "@npm//protobufjs",
-        "@npm//react",
     ],
 )