--- a/app/service/BUILD.bazel
+++ b/app/service/BUILD.bazel
@@ -7,11 +7,11 @@ ts_library(
     srcs = glob(["*.ts"]),
     strict = True,
     deps = [
+        "//:node_modules/oidc-client",
+        "//:node_modules/protobufjs",
+        "//:node_modules/rxjs",
         "//app/util:async",
         "//proto:buildbuddy_service_ts_proto",
         "//proto:context_ts_proto",
-        "@npm//oidc-client",
-        "@npm//protobufjs",
-        "@npm//rxjs",
     ],
 )