--- a/enterprise/app/BUILD.bazel
+++ b/enterprise/app/BUILD.bazel
@@ -1,5 +1,5 @@
+load("@aspect_rules_esbuild//esbuild:defs.bzl", "esbuild")
 load("//rules/typescript:index.bzl", "ts_library")
-load("@npm//@bazel/esbuild:index.bzl", "esbuild", "esbuild_config")
 load("//rules/sha:index.bzl", "sha")
 
 package(default_visibility = ["//enterprise:__subpackages__"])
@@ -9,24 +9,9 @@ config_setting(
     values = {"compilation_mode": "fastbuild"},
 )
 
-esbuild_config(
-    name = "esbuild_config",
-    config_file = "esbuild.config.mjs",
-)
-
 esbuild(
     name = "app_bundle",
-    args = {
-        "resolveExtensions": [
-            ".mjs",
-            ".js",
-        ],
-        "loader": {
-            ".ttf": "binary",
-            ".css": "binary",
-        },
-    },
-    config = ":esbuild_config",
+    config = ":esbuild.config.mjs",
     define = {"process.env.NODE_ENV": '"production"'},
     entry_points = [
         "app.tsx",
@@ -39,6 +24,7 @@ esbuild(
     visibility = ["//visibility:public"],
     deps = [
         ":enterprise_app",
+        "//:node_modules/path-browserify",
     ],
 )
 
@@ -46,11 +32,11 @@ ts_library(
     name = "enterprise_app",
     srcs = glob(["*.tsx"]),
     deps = [
+        "//:node_modules/@types/react",
+        "//:node_modules/@types/react-dom",
+        "//:node_modules/react",
+        "//:node_modules/react-dom",
         "//enterprise/app/root",
-        "@npm//@types/react",
-        "@npm//@types/react-dom",
-        "@npm//react",
-        "@npm//react-dom",
     ],
 )