--- a/enterprise/app/org/BUILD
+++ b/enterprise/app/org/BUILD
@@ -9,6 +9,9 @@ ts_library(
     srcs = glob(["*.tsx"]),
     strict = True,
     deps = [
+        "//:node_modules/@types/react",
+        "//:node_modules/lucide-react",
+        "//:node_modules/react",
         "//app/alert",
         "//app/auth",
         "//app/capabilities",
@@ -24,8 +27,5 @@ ts_library(
         "//app/util:errors",
         "//proto:group_ts_proto",
         "//proto:user_id_ts_proto",
-        "@npm//@types/react",
-        "@npm//lucide-react",
-        "@npm//react",
     ],
 )