--- a/enterprise/app/root/BUILD.bazel
+++ b/enterprise/app/root/BUILD.bazel
@@ -8,6 +8,11 @@ ts_library(
     name = "root",
     srcs = glob(["*.tsx"]),
     deps = [
+        "//:node_modules/@types/node",
+        "//:node_modules/@types/react",
+        "//:node_modules/lucide-react",
+        "//:node_modules/protobufjs",
+        "//:node_modules/react",
         "//app/alert",
         "//app/auth",
         "//app/capabilities",
@@ -35,10 +40,5 @@ ts_library(
         "//enterprise/app/usage",
         "//enterprise/app/workflows",
         "//proto:user_ts_proto",
-        "@npm//@types/node",
-        "@npm//@types/react",
-        "@npm//lucide-react",
-        "@npm//protobufjs",
-        "@npm//react",
     ],
 )