… .prettierignore
… app/compare/BUILD
… app/invocation/BUILD
… app/terminal/BUILD
… docs/rbe-platforms.md
… enterprise/app/code/BUILD
… enterprise/app/executors/BUILD
… enterprise/app/filter/BUILD
… enterprise/app/history/BUILD
… enterprise/app/tap/BUILD
… enterprise/app/trends/BUILD
… enterprise/app/workflows/BUILD
… package.json
… pnpm-lock.yaml
… proto/defs.bzl
… rules/typescript/index.bzl
… server/metrics/BUILD
… tools/prettier/BUILD
… tools/prettier/prettier.sh
… tsconfig.json
… yarn.lock