Rebases @thesayyn's PR https://github.com/buildbuddy-io/buildbuddy/pull/2143

Tested:
- [x] Can build and run OSS app
- [x] Can build and run enterprise app
- [x] Can build and run server image from internal repo
  - [x] Depends on https://github.com/aspect-build/rules_swc/pull/128
  - [x] Depends on https://github.com/aspect-build/rules_esbuild/pull/108
- [ ] Make sure disk usage is acceptable - the new rules do a lot of copying, and we might need to increase the disk space available to workflows to accommodate this, or see if there is a way to use less disk space, e.g. hardlinking 
- [ ] Should not break website deploy rules 
- [ ] Test dev push to make sure there is no app bundle SHA mismatch
- [x] RBE Performance: `bazel build //enterprise/app:app_bundle --config=remote-dev`: LGTM; improved by about 2s (15s -> 13s, N=100)
  - Update: This should now be even faster than the results shown above, since rules_swc was updated to use a rust-based CLI (swcx) rather than using a nodejs wrapper around swc. 
- [x] RBE Performance: `bazel test //app/... //enterprise/app/... --config=remote-dev --nocache_test_results`: LGTM; improved by about 0.6s (7.5s -> 6.9s, N=100)

---

**Version bump**: Patch <!-- Required. Choose from: Major, Minor, Patch, None -->

<!-- See https://semver.org/#semantic-versioning-specification-semver. Summary:
* Major: Breaking change that causes existing functionality to not work as expected.
* Minor: Non-breaking change that adds functionality (examples: new feature; new API options)
* Patch: Non-breaking change that fixes an issue, improves performance, or refactors
         code.
* None:  Changed files are not included in releases (tests, docs, development setup,
         production configs)
-->

<!-- Optional:
**Related issues**: Fixes #1, Unblocks #2 ...
-->