@bduffany 2022-12-14
Revisiting now that we're on Bazel 6.0