@bduffany 2023-01-05
FYI I am just waiting on https://github.com/aspect-build/rules_swc/pull/128/files and https://github.com/aspect-build/rules_esbuild/pull/108 to get merged upstream.