@sluongng 2023-01-27
It seems like for this to work, we gonna need to migrate `//website` from yarn to `pnpm`, then use the `@pnpm//:pnpm` binary in replacement for `@yarn//:yarn` binary 🤔 

I found this which could be helpful https://dev.to/andreychernykh/yarn-npm-to-pnpm-migration-guide-2n04