@sluongng 2023-02-24
Example on how to migrate docusaurus targets https://github.com/formatjs/formatjs/blob/2d7bd516d7e6fc85d3e208bac390d81396d43d17/website/BUILD